Fix compatibility, use overload with properties
